Fixed a bunch of exploits and glitches that could be made with decimal numbers, due to the fact that i forgot about them. Added health withdraw. Cleaned some code

+++ src/main/java/me/imlukas/withdrawer/managers/Note.java
@@ -18,7 +18,7 @@ public class Note extends Manager {
         }
         if (economyUtil.hasMoney(player, money * amount)) {
             economyUtil.removeMoney(player, total);
-            ItemStack noteItem = setItemProperties(player, (int) money);
+            ItemStack noteItem = setItemProperties(player, money);
             if (amount > 1) {
                 for (int i = 0; i < amount; i++) {
                     player.getInventory().addItem(noteItem);
